+++ blosc2/__init__.py
@@ -72,6 +72,7 @@ from .core import (
     get_cbuffer_sizes,
     load_array,
     load_tensor,
+    open,
     pack,
     pack_array,
     pack_array2,
@@ -93,7 +94,7 @@ from .core import (
     unpack_tensor,
 )
 from .ndarray import NDArray, asarray, copy, empty, from_buffer, full, zeros
-from .SChunk import SChunk, open
+from .SChunk import SChunk
 from .version import __version__
 
 # Registry for postfilters
