+++ doc/reference/index.rst
@@ -6,4 +6,5 @@ API Reference
 
     top_level
     schunk_api
+    ndarray_api
     utils
