+++ requirements-runtime.txt
@@ -2,3 +2,4 @@ msgpack
 py-cpuinfo
 ndindex
 ndindex>=1.4
+numpy>=1.20.3
